refactor: update import paths to use consistent naming convention for Firebase modules
