+++ src/components/modules/auth/server/authentication.firebase.ts
