+++ src/components/modules/setting/hooks/api-keys.hook.ts
@@ -2,7 +2,7 @@ import { useGlobalAuthenticationStore } from "@/core/store/data";
 import { useChangeUtils } from "@/utils/hook/input-visibility.hook";
 import { useState } from "react";
 import { requestApiKey } from "../services/requestApiKey";
-import { getUser } from "../../auth/server/authentication-firebase";
+import { getUser } from "../../auth/server/authentication.firebase";
 import { removeApiKey } from "../server/api-key-firebase";
 import { toast } from "@/hooks/use-toast";
